fix:Biginteger로바꾸기

+++ src/main/java/clutch/clutchserver/calculate/dto/CalculateRequestDto.java
@@ -2,12 +2,14 @@ package clutch.clutchserver.calculate.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class CalculateRequestDto {
 
     private Long buildingId;
 
-    private int deposit;
+    private BigInteger deposit;
 
     private Boolean isDangerous;
