+++ src/main/java/clutch/clutchserver/calculate/dto/CalculateResponseDto.java
@@ -3,6 +3,8 @@ package clutch.clutchserver.calculate.dto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 
+import java.math.BigInteger;
+
 public class CalculateResponseDto {
 
     @Schema(description="계산 ID",example= "1")
@@ -12,13 +14,13 @@ public class CalculateResponseDto {
     private Long buildingId;
 
     @Schema(description="전세금",example= "1")
-    private int deposit;
+    private BigInteger deposit;
 
     @Schema(description="위험 여부",example= "1")
     private Boolean isDangerous;
 
     @Builder
-    public CalculateResponseDto(Long id, Long buildingId, int deposit, Boolean isDangerous) {
+    public CalculateResponseDto(Long id, Long buildingId, BigInteger deposit, Boolean isDangerous) {
         this.id = id;
         this.buildingId = buildingId;
         this.deposit = deposit;
